moved the encoding of the data to another file. Because i wanted to seperate the concerns
