+++ src/train.py
@@ -1,57 +1,13 @@
-from transformers import AutoTokenizer, AutoModelForSequenceClassification, TrainingArguments, Trainer
-from sklearn.model_selection import train_test_split
+from transformers import AutoModelForSequenceClassification, TrainingArguments, Trainer
 import torch
 import pandas as pd
-import datasets
 import os, multiprocessing
+from encode_data import encode_data, TOKENIZER
 
 TRAIN_MODEL = True
 MODEL_NAME = "bert-base-uncased"
-CONTEXT_LENGTH = 512
 NUM_PROC = min(100, multiprocessing.cpu_count() - 1)
 FRAC_OF_DATA = 1
-
-tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
-
-def get_issues_df():
-    issues_df = pd.read_json(os.path.join("data", "issues.json"))
-    issues_df['label'] = issues_df['assignee'].astype('category').cat.codes
-    if FRAC_OF_DATA < 1:
-        issues_df = issues_df.sample(frac=FRAC_OF_DATA, random_state=1, weights="label").reset_index(drop=True)
-    return issues_df
-
-def tokenize(examples):
-    encoding  = tokenizer(examples["text"], padding="max_length", truncation=True, max_length=CONTEXT_LENGTH)
-    encoding["label"] = examples["label"]
-    return encoding
-
-def encode_dataset(force=False, filename=os.path.join("data", f"encoded_dataset_{int(FRAC_OF_DATA*100):03d}")):
-    if not force and os.path.exists(filename):
-        return datasets.load_from_disk(filename)
-    print(f"Saving encoded dataset to {filename}")
-
-    issues_df = get_issues_df()
-
-    train_df = issues_df[issues_df["github_id"] <= 210_000]
-    train_df, eval_df = train_test_split(train_df, test_size=0.1, random_state=42, stratify=train_df["label"])
-    test_df = issues_df[(210_000 < issues_df["github_id"]) & (issues_df["github_id"] <= 220_000)]
-
-    dataset = datasets.DatasetDict({
-        "train": datasets.Dataset.from_pandas(train_df),
-        "eval": datasets.Dataset.from_pandas(eval_df),
-        "test": datasets.Dataset.from_pandas(test_df),
-    })
-    # dataset = dataset.filter(lambda x: len(tokenizer(x["cleaned_body"])["input_ids"]) <= CONTEXT_LENGTH, num_proc=NUM_PROC)
-    
-    encoded_dataset = dataset.map(
-        tokenize,
-        batched=True,
-        remove_columns=[column for column in dataset['train'].column_names if column not in ["label", "github_id"]], 
-        # remove_columns=dataset['train'].column_names,
-        num_proc=NUM_PROC
-    )
-    encoded_dataset.save_to_disk(filename)
-    return encoded_dataset
 
 def train_model(model, dataset, output_dir=os.path.join("data", "checkpoints")):
     os.environ["CUDA_VISIBLE_DEVICES"] = "2"
@@ -73,7 +29,7 @@ def train_model(model, dataset, output_dir=os.path.join("data", "checkpoints")):
         args,
         train_dataset=dataset["train"],
         eval_dataset=dataset["eval"],
-        tokenizer=tokenizer,
+        tokenizer=TOKENIZER,
         # compute_metrics=compute_metrics
     )
 
@@ -89,7 +45,7 @@ def train_model(model, dataset, output_dir=os.path.join("data", "checkpoints")):
     return trainer
 
 def main():
-    encoded_dataset = encode_dataset()
+    encoded_dataset = encode_data()
     encoded_dataset.set_format("torch")
 
     labels = set(map(lambda x: int(x), encoded_dataset["train"]["label"]))
